+++ src/crate/theme/rtd/conf/dbal.py
@@ -21,8 +21,8 @@
 
 from crate.theme.rtd.conf import *
 
-project = u'CrateDB JDBC'
+project = u'DBAL'
 
 html_theme_options.update({
-    'canonical_url_path': 'docs/clients/jdbc/en/latest/',
+    'canonical_url_path': 'docs/clients/dbal/en/latest/',
 })