+++ src/crate/theme/rtd/conf/npgsql.py
@@ -21,8 +21,8 @@
 
 from crate.theme.rtd.conf import *
 
-project = u'CrateDB Python'
+project = u'Npgsql'
 
 html_theme_options.update({
-    'canonical_url_path': 'docs/clients/python/en/latest/',
+    'canonical_url_path': 'docs/clients/npgsql/en/latest/',
 })